requirements.txt
streamlit>=1.30.0
flask>=2.3.0
google-generativeai>=0.3.2
googletrans==4.0.0-rc1
gTTS>=2.3.2
SpeechRecognition>=3.10.0
pyaudio>=0.2.13
httpx==0.13.3
